updated the GUI loop. some bug fixes to scheduler
i tried to change the GUI to display the ready and waiting queues, but i don't know much about the UI classes. it's showing the waiting queue (left) having a bunch of extra processes, but i checked the debugger and there's nothing there. i think its the UI not updating correctly.