fixed the waiting queue not waiting. found a weird bug
so if you load in a bunch of test files and start executing you can see that they all have the same commands, obviously. however if you then load in an ide, all of the programs get the command list of the ide. even if the ide goes into the waiting queue. 

the test program only only has CALCULATE 33 in it, and it starts getting ones much higher than 33 and also OUT commands, which it never had.